Updated :tag_id index removal to also remove by index name

+++ db/migrate/20150724142715_add_missing_unique_indices.acts_as_taggable_on_engine.rb
@@ -3,7 +3,10 @@ class AddMissingUniqueIndices < ActiveRecord::Migration
   def self.up
     add_index :tags, :name, unique: true
 
-    remove_index :taggings, :tag_id
+    remove_index :taggings, :tag_id if index_exists?(:taggings, :tag_id)
+    # in case the previous line does nothing, also try removing it by "index name"
+    remove_index(:taggings, name: "index_taggings_on_tag_id") if index_exists?(:taggings, :tag_id, name: "index_taggings_on_tag_id")
+    
     remove_index :taggings, [:taggable_id, :taggable_type, :context]
     add_index :taggings,
               [:tag_id, :taggable_id, :taggable_type, :context, :tagger_id, :tagger_type],
